mammoth update, hope nothing's broken
let's see..
reworked the flow in main. now it asks for the passwords to all the accounts before trying to open any of the feed. this is more of a usability fix.
the code now parses all the accounts (both supplied with --username flags and as unnamed arguments) with the same code
if the password is wrong it ignores that account, moving onto the next
a lot of the commented code (but not all) is from a failed implementation of max retries for the password. i'll be working on this later
added some exceptions to make the above changes work
fixed a bug where FancyURLopenerMod just assumed username and password were being passed to it. stupid

i really really need to get the hang of making incremental commits. spent nearly an hour removing code from a failed bugfix